+++ core/OS/Search/Profile.cpp
@@ -183,8 +183,8 @@ namespace OS {
 					error = EProfileResponseType_Success;
 					json_t *profiles_obj = json_object_get(json_data, "profiles");
 					if (profiles_obj) {
-						int num_profiles = json_array_size(profiles_obj);
-						for (int i = 0; i < num_profiles; i++) {
+						size_t num_profiles = json_array_size(profiles_obj);
+						for (size_t i = 0; i < num_profiles; i++) {
 							json_t *profile_obj = json_array_get(profiles_obj, i);
 							OS::Profile profile = OS::LoadProfileFromJson(profile_obj);
 							if (users_map.find(profile.userid) == users_map.end()) {
